refactor summary page, added notes
